# infra/api/auth.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Para Auth API",
    "version": "0.5.2",
...
  },
  "servers": [
    {
      "url": "https://nobeta.cn/auth",
      "description": "默认公开 issuer 地址；需以实际部署 AUTH_ISSUER 为准。"
    },
    {
      "url": "http://localhost:9000/auth",
      "description": "本地开发示例：需 AUTH_ALLOW_LOCAL_HTTP=true、AUTH_ISSUER=http://localhost:9000/auth、AUTH_SECURE_COOKIES=false；不能据此认为已部署。"
    }
  ],
...
  "security": [],
  "paths": {
    "/api/csrf": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "getCsrfToken",
        "summary": "获取会话 CSRF 令牌",
        "description": "匿名与已登录用户均可调用。读取令牌会按需创建会话；保存响应 Cookie，再使用返回 token。登录成功后会话 ID/CSRF 可能变化，重新获取。多次读取的掩码 token 字符串可能不同，不能按字符串相等判断会话变化。\n\n### cURL 调用示例\n\n```bash\ncurl --include \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  \"$AUTH_BASE/api/csrf\"\n```\n\n变量说明见文档开头。",
        "security": [],
        "responses": {
          "200": {
            "description": "CSRF 参数与令牌。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfView"
                },
                "examples": {
                  "CsrfToken": {
                    "$ref": "#/components/examples/CsrfToken"
                  }
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        },
        "x-codeSamples": [
          {
            "lang": "Shell",
            "label": "cURL",
            "source": "curl --include \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  \"$AUTH_BASE/api/csrf\""
          }
        ],
        "x-implementation-source": [
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountController.java",
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountService.java"
        ]
      }
    },
    "/api/accounts/register": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "operationId": "registerAccount",
        "summary": "注册账号",
        "description": "注册开关默认启用；关闭时 403。用户名存储为小写，大小写变体共享唯一约束。创建 enabled=true 的普通账号，不赋予管理员角色，不建立登录身份，不签发 Token。匿名写请求仍需先获取 CSRF 与会话 Cookie。\n\n### 业务失败情形\n\n| 状态 | 触发条件 | 源码 reason |\n| --- | --- | --- |\n| 400 | 密码不足 12 个 UTF-16 代码单元或超过 72 UTF-8 字节 | `Password must contain at least 12 characters and at most 72 UTF-8 bytes` |\n| 403 | 注册开关关闭 | `Registration is disabled` |\n| 409 | 规范化后用户名已存在 | `The identifier already exists` |\n\n业务 reason 用于解释源码行为，不保证作为 response.message/detail 输出。\n\n### cURL 调用示例\n\n```bash\ncurl --include \\\n  --request POST \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --header \"X-CSRF-TOKEN: $CSRF_TOKEN\" \\\n  --header \"Content-Type: application/json\" \\\n  --data '{\"username\":\"Alice\",\"password\":\"example-passphrase-2026\"}' \\\n  \"$AUTH_BASE/api/accounts/register\"\n```\n\n变量说明见文档开头。",
        "security": [],
        "responses": {
          "201": {
            "description": "已创建账号；不返回 Location，也不自动登录。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountView"
                },
                "examples": {
                  "AccountEnabled": {
                    "$ref": "#/components/examples/AccountEnabled"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BusinessBadRequest"
          },
          "403": {
            "description": "CSRF 失败时 application/problem+json；Registration is disabled 由业务 ResponseStatusException 返回，可能 application/json。",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootError"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "examples": {
                "RegisterAccount": {
                  "$ref": "#/components/examples/RegisterAccount"
                }
              }
            }
          }
        },
        "x-business-error-cases": [
          {
            "status": 400,
            "cause": "密码不足 12 个 UTF-16 代码单元或超过 72 UTF-8 字节",
            "sourceReason": "Password must contain at least 12 characters and at most 72 UTF-8 bytes",
            "reasonGuaranteedInResponse": false
          },
          {
            "status": 403,
            "cause": "注册开关关闭",
            "sourceReason": "Registration is disabled",
            "reasonGuaranteedInResponse": false
          },
          {
            "status": 409,
            "cause": "规范化后用户名已存在",
            "sourceReason": "The identifier already exists",
            "reasonGuaranteedInResponse": false
          }
        ],
        "x-codeSamples": [
          {
            "lang": "Shell",
            "label": "cURL",
            "source": "curl --include \\\n  --request POST \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --header \"X-CSRF-TOKEN: $CSRF_TOKEN\" \\\n  --header \"Content-Type: application/json\" \\\n  --data '{\"username\":\"Alice\",\"password\":\"example-passphrase-2026\"}' \\\n  \"$AUTH_BASE/api/accounts/register\""
          }
        ],
        "x-implementation-source": [
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountController.java",
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountService.java"
        ]
      }
    },
    "/api/accounts/me": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "operationId": "getCurrentAccount",
        "summary": "查询当前账号",
        "description": "从会话中的稳定 subject 查询账号。每次已有会话请求会刷新账号状态和管理员权限，停用或不存在的账号会话将失效。Access Token 不用于该业务接口。\n\n### cURL 调用示例\n\n```bash\ncurl --include \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  \"$AUTH_BASE/api/accounts/me\"\n```\n\n变量说明见文档开头。",
        "security": [
          {
            "AuthSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "当前账号。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountView"
                },
                "examples": {
                  "AccountEnabled": {
                    "$ref": "#/components/examples/AccountEnabled"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BusinessBadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "x-codeSamples": [
          {
            "lang": "Shell",
            "label": "cURL",
            "source": "curl --include \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  \"$AUTH_BASE/api/accounts/me\""
          }
        ],
        "x-implementation-source": [
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountController.java",
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountService.java"
        ]
      }
    },
    "/api/accounts/me/password": {
      "put": {
        "tags": [
          "Accounts"
        ],
        "operationId": "changeCurrentPassword",
        "summary": "修改当前账号密码",
        "description": "验证 currentPassword；使用比较旧哈希的并发更新，冲突返回 409。成功删除该 subject 全部授权记录及 Redis 会话，并退出当前请求会话。调用者需要重新登录。不会主动撤销外部应用本地会话或使仅离线验签的旧 JWT 立即失效。\n\n### 业务失败情形\n\n| 状态 | 触发条件 | 源码 reason |\n| --- | --- | --- |\n| 400 | 当前密码错误 | `Current password is incorrect` |\n| 409 | 读取旧哈希后，凭据被其他请求更新 | `Credentials changed concurrently` |\n| 404 | 主体对应账号不存在 | `Account not found` |\n\n业务 reason 用于解释源码行为，不保证作为 response.message/detail 输出。\n\n### cURL 调用示例\n\n```bash\ncurl --include \\\n  --request PUT \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --header \"X-CSRF-TOKEN: $CSRF_TOKEN\" \\\n  --header \"Content-Type: application/json\" \\\n  --data '{\"currentPassword\":\"example-passphrase-2026\",\"newPassword\":\"replacement-passphrase-2026\"}' \\\n  \"$AUTH_BASE/api/accounts/me/password\"\n```\n\n变量说明见文档开头。",
        "security": [
          {
            "AuthSession": []
          }
        ],
        "responses": {
          "204": {
            "description": "成功，无响应体；当前会话结束。",
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BusinessBadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRequest"
              },
              "examples": {
                "ChangePassword": {
                  "$ref": "#/components/examples/ChangePassword"
                }
              }
            }
          }
        },
        "x-business-error-cases": [
          {
            "status": 400,
            "cause": "当前密码错误",
            "sourceReason": "Current password is incorrect",
            "reasonGuaranteedInResponse": false
          },
          {
            "status": 409,
            "cause": "读取旧哈希后，凭据被其他请求更新",
            "sourceReason": "Credentials changed concurrently",
            "reasonGuaranteedInResponse": false
          },
          {
            "status": 404,
            "cause": "主体对应账号不存在",
            "sourceReason": "Account not found",
            "reasonGuaranteedInResponse": false
          }
        ],
        "x-codeSamples": [
          {
            "lang": "Shell",
            "label": "cURL",
            "source": "curl --include \\\n  --request PUT \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --header \"X-CSRF-TOKEN: $CSRF_TOKEN\" \\\n  --header \"Content-Type: application/json\" \\\n  --data '{\"currentPassword\":\"example-passphrase-2026\",\"newPassword\":\"replacement-passphrase-2026\"}' \\\n  \"$AUTH_BASE/api/accounts/me/password\""
          }
        ],
        "x-implementation-source": [
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountController.java",
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountService.java"
        ]
      }
    },
    "/api/admin/accounts": {
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "listAccounts",
        "summary": "分页查询账号",
        "description": "要求会话具备 ROLE_AUTH_ADMIN。按数据库 id 升序返回；没有用户名搜索或 enabled 筛选参数。\n\n### cURL 调用示例\n\n```bash\ncurl --include \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --get \\\n  --data-urlencode \"page=1\" \\\n  --data-urlencode \"size=20\" \\\n  \"$AUTH_BASE/api/admin/accounts\"\n```\n\n变量说明见文档开头。",
        "security": [
          {
            "AuthSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "账号分页。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPage"
                },
                "examples": {
                  "AccountsPage": {
                    "$ref": "#/components/examples/AccountsPage"
                  },
                  "AccountsEmptyPage": {
                    "$ref": "#/components/examples/AccountsEmptyPage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BusinessBadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          }
        ],
        "x-required-authorities": [
          "ROLE_AUTH_ADMIN"
        ],
        "x-codeSamples": [
          {
            "lang": "Shell",
            "label": "cURL",
            "source": "curl --include \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --get \\\n  --data-urlencode \"page=1\" \\\n  --data-urlencode \"size=20\" \\\n  \"$AUTH_BASE/api/admin/accounts\""
          }
        ],
        "x-implementation-source": [
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountController.java",
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountService.java"
        ]
      }
    },
    "/api/admin/accounts/{id}/status": {
      "put": {
        "tags": [
          "Administration"
        ],
        "operationId": "setAccountStatus",
        "summary": "启用或停用账号",
        "description": "要求 ROLE_AUTH_ADMIN。不能停用正在执行操作的管理员自己（409）。停用删除该账号全部授权记录和 Redis 会话；重新启用不恢复已删除记录。重复设置仍执行数据库更新；重复停用仍执行撤销。\n\n### 业务失败情形\n\n| 状态 | 触发条件 | 源码 reason |\n| --- | --- | --- |\n| 404 | 账号不存在 | `Account not found` |\n| 409 | 管理员停用当前自己的账号 | `An administrator cannot disable the active account` |\n\n业务 reason 用于解释源码行为，不保证作为 response.message/detail 输出。\n\n### cURL 调用示例\n\n```bash\ncurl --include \\\n  --request PUT \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --header \"X-CSRF-TOKEN: $CSRF_TOKEN\" \\\n  --header \"Content-Type: application/json\" \\\n  --data '{\"enabled\":false}' \\\n  \"$AUTH_BASE/api/admin/accounts/$ACCOUNT_ID/status\"\n```\n\n变量说明见文档开头。",
        "security": [
          {
            "AuthSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "更新后的账号。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountView"
                },
                "examples": {
                  "AccountEnabled": {
                    "$ref": "#/components/examples/AccountEnabled"
                  },
                  "AccountDisabled": {
                    "$ref": "#/components/examples/AccountDisabled"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BusinessBadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedError"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              },
              "examples": {
                "Enable": {
                  "$ref": "#/components/examples/Enable"
                },
                "Disable": {
                  "$ref": "#/components/examples/Disable"
                }
              }
            }
          }
        },
        "x-required-authorities": [
          "ROLE_AUTH_ADMIN"
        ],
        "x-business-error-cases": [
          {
            "status": 404,
            "cause": "账号不存在",
            "sourceReason": "Account not found",
            "reasonGuaranteedInResponse": false
          },
          {
            "status": 409,
            "cause": "管理员停用当前自己的账号",
            "sourceReason": "An administrator cannot disable the active account",
            "reasonGuaranteedInResponse": false
          }
        ],
        "x-codeSamples": [
          {
            "lang": "Shell",
            "label": "cURL",
            "source": "curl --include \\\n  --request PUT \\\n  --cookie auth.cookies --cookie-jar auth.cookies \\\n  --header \"X-CSRF-TOKEN: $CSRF_TOKEN\" \\\n  --header \"Content-Type: application/json\" \\\n  --data '{\"enabled\":false}' \\\n  \"$AUTH_BASE/api/admin/accounts/$ACCOUNT_ID/status\""
          }
        ],
        "x-implementation-source": [
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountController.java",
          "backend/auth/src/main/java/cn/nobeta/auth/module/account/AccountService.java"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "AUTH_SESSION",
        "description": "登录后建立的 Redis Session。HttpOnly，前端不直接读取 Cookie。浏览器同源或凭据请求发送 Cookie；当前后端没有配置跨域 CORS 支持。默认 30 分钟会话超时。"
      },
      "ClientSecretBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "OAuth client_secret_basic：公开 clientId 与明文 clientSecret，不是账号 username/password。只用于协议后端端点；密钥保存在可信后端。"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "令牌端点签发的有效 Access Token，UserInfo 还要求有效 openid 授权。不可使用 ID Token；不可代替业务 Cookie。"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "X-CSRF-TOKEN",
        "in": "header",
        "required": true,
        "description": "取 GET /api/csrf 的 token，名称以 headerName 实际值为准。JSON 写接口使用请求头；不接受仅发送 Cookie 代替令牌。",
        "schema": {
          "type": "string"
        },
        "example": "csrf-token-placeholder"
      },
...
    },
    "headers": {
...
      "SetCookie": {
        "description": "可能创建、轮换或清除 AUTH_SESSION。HttpOnly; Path=/auth; SameSite=Lax；Secure 默认 true。具体编码及重复头由会话组件决定。",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "缺少有效 AUTH_SESSION，401；Bearer 不能替代业务会话认证。",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "NoSession": {
                "$ref": "#/components/examples/NoSession"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "缺少 ROLE_AUTH_ADMIN，或 CSRF 令牌缺失/错误，403。未登录的写请求也可能先在 CSRF 阶段返回 403。",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "AccessDenied": {
                "$ref": "#/components/examples/AccessDenied"
              }
            }
          }
        }
      },
      "BusinessBadRequest": {
        "description": "400：校验错误为 ProblemDetail；业务 ResponseStatusException 或参数转换错误可能走 Boot /error。不可依赖 detail/message 一定存在。",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "InvalidParameters": {
                "$ref": "#/components/examples/InvalidParameters"
              }
            }
          },
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BootError"
            },
            "examples": {
              "Boot400": {
                "$ref": "#/components/examples/Boot400"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "404：Account not found 或 Client not found。reason 是源码语义，默认不保证作为响应字段公开。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BootError"
            },
            "examples": {
              "Boot404": {
                "$ref": "#/components/examples/Boot404"
              }
            }
          },
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "Problem404": {
                "$ref": "#/components/examples/Problem404"
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "409：标识重复、禁止停用自己的账号或凭证并发更新。DuplicateKeyException 显式返回 ProblemDetail，其他业务冲突可能走 Boot /error。",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "IdentifierConflict": {
                "$ref": "#/components/examples/IdentifierConflict"
              }
            }
          },
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BootError"
            },
            "examples": {
              "Boot409": {
                "$ref": "#/components/examples/Boot409"
              }
            }
          }
        }
      },
...
      "UnexpectedError": {
        "description": "500：未处理的数据库、Redis、序列化等内部异常。当前业务无统一 500 DTO，可能经 Boot /error；不承诺泄露内部异常详情。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BootError"
            },
            "examples": {
              "Boot500": {
                "$ref": "#/components/examples/Boot500"
              }
            }
          },
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "examples": {
              "Problem500": {
                "$ref": "#/components/examples/Problem500"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "CsrfView": {
        "type": "object",
        "required": [
          "headerName",
          "parameterName",
          "token"
        ],
        "properties": {
          "headerName": {
            "type": "string",
            "description": "将 token 放入此请求头；框架默认值为 X-CSRF-TOKEN。",
            "example": "X-CSRF-TOKEN",
            "readOnly": true
          },
          "parameterName": {
            "type": "string",
            "description": "表单参数名；框架默认值为 _csrf。",
            "example": "_csrf",
            "readOnly": true
          },
          "token": {
            "type": "string",
            "description": "与当前会话绑定的 CSRF 防伪令牌；不是访问令牌，不能跨会话复用。",
            "minLength": 1,
            "readOnly": true
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "注册请求先通过 Bean Validation，故前后空格会被正则拒绝；服务存储时转为小写。登录会 strip 并转为小写。",
            "minLength": 3,
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "example": "Alice"
          },
          "password": {
            "type": "string",
            "description": "非空白；Java String.length 至少 12，UTF-8 编码最多 72 字节。maxLength 不能表达 UTF-8 字节约束。",
            "minLength": 12,
            "maxLength": 72,
            "format": "password",
            "writeOnly": true,
            "x-utf8-max-bytes": 72,
            "x-java-min-utf16-code-units": 12
          }
        },
        "description": "\n\n### 注册与登录\n\n注册用户名长度 `3..64`，匹配 `[A-Za-z0-9][A-Za-z0-9._-]*`；存储为小写。注册 DTO 的正则先执行，因此带前后空格的用户名会失败。登录加载账号时则会去前后空白并转小写。\n\n新密码非空白，Java `String.length()` 至少 12，UTF-8 最多 72 字节。Bean Validation 还限制字符长度最多 72。ASCII 密码可简单理解为 12..72 个字符；中文和其他 Unicode 字符按 UTF-8 字节计算，不能仅靠前端 `maxLength=72` 确认有效。当前密码仅要求非空白，以兼容历史密码。\n\n`AUTH_REGISTRATION_ENABLED` 默认 `true`。注册成功返回 `201 AccountView`，账号启用但没有管理员角色，不自动登录或签发令牌。重复用户名，包括大小写规范化后相同的名称，返回 `409`。\n\n框架登录只接受 `application/x-www-form-urlencoded`：`username`、`password`、`_csrf`。没有 JSON `POST /api/login` 或 `/api/accounts/login`。\n\n登录失败默认 `302 /auth/login?error`。成功跳转已保存的浏览器请求或默认 `/`。当前安全规则拒绝未明确允许的路径，因此直接登录后不能假设 `/auth/` 是有效首页；SSO 从授权入口发起更符合当前流程。"
      },
      "PasswordRequest": {
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "当前密码；仅要求非空白，不强加新密码长度规则，以兼容历史密码。",
            "minLength": 1,
            "format": "password",
            "writeOnly": true
          },
          "newPassword": {
            "type": "string",
            "description": "非空白；至少 12 个 Java UTF-16 代码单元，最多 72 UTF-8 字节。",
            "minLength": 12,
            "maxLength": 72,
            "format": "password",
            "writeOnly": true,
            "x-utf8-max-bytes": 72,
            "x-java-min-utf16-code-units": 12
          }
        }
      },
      "AccountView": {
        "type": "object",
        "description": "所有账号响应均不返回密码、密码哈希或角色列表。\n\n### 时间与身份字段\n\n`AccountView` 返回 `id`、`subject`、`username`、`enabled`、`createTime`、`updateTime`，不返回密码哈希或角色。\n\n数据库连接配置以 UTC 为时间基准。账号时间字段类型是 `LocalDateTime`，序列化示例 `2026-10-03T02:00:00`，没有 `Z` 或时区偏移，因此规范没有把它错误声明为带偏移的 RFC 3339 `date-time`。客户端应按 UTC 解释。客户端配置中的 `clientIdIssuedAt`、`clientSecretExpiresAt` 是 `Instant`，使用带偏移的时间，例如 `2026-10-03T02:00:00Z`。",
        "required": [
          "id",
          "subject",
          "username",
          "enabled",
          "createTime",
          "updateTime"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "数据库内部账号 ID；不等于 OIDC sub。",
            "format": "int64",
            "minimum": 1,
            "readOnly": true
          },
          "subject": {
            "type": "string",
            "description": "稳定身份 UUID；OIDC sub 和会话 principal 使用此值。应用应按 (issuer, sub) 关联用户。",
            "format": "uuid",
            "readOnly": true
          },
          "username": {
            "type": "string",
            "description": "规范化为小写的登录账号。",
            "minLength": 3,
            "maxLength": 64,
            "readOnly": true
          },
          "enabled": {
            "type": "boolean",
            "description": "status == 1 时为 true。",
            "readOnly": true
          },
          "createTime": {
            "type": "string",
            "description": "数据库 UTC 时间，Java LocalDateTime 序列化不带 Z 或偏移；不可按本地时区误读。",
            "example": "2026-10-03T02:00:00",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,9})?$",
            "readOnly": true
          },
          "updateTime": {
            "type": "string",
            "description": "同 createTime；无时区后缀的 UTC LocalDateTime。",
            "example": "2026-10-03T02:00:00",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,9})?$",
            "readOnly": true
          }
        },
        "example": {
          "id": 42,
          "subject": "7d8a8d55-e9cb-4db4-a945-6f4c21cbab31",
          "username": "alice",
          "enabled": true,
          "createTime": "2026-10-03T02:00:00",
          "updateTime": "2026-10-03T02:00:00"
        }
      },
...
      "AccountPage": {
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountView"
            },
            "description": "当前页记录，空页返回 []。"
          },
          "total": {
            "type": "integer",
            "description": "符合条件的总记录数，不是总页数。",
            "format": "int64",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "description": "从 1 开始的请求页码，超出总页数不会自动修正。",
            "minimum": 1,
            "maximum": 1000000
          },
          "size": {
            "type": "integer",
            "description": "请求页容量，不等于 items 实际长度。",
            "minimum": 1,
            "maximum": 100
          }
        },
        "example": {
          "items": [
            {
              "id": 42,
              "subject": "7d8a8d55-e9cb-4db4-a945-6f4c21cbab31",
              "username": "alice",
              "enabled": true,
              "createTime": "2026-10-03T02:00:00",
              "updateTime": "2026-10-03T02:00:00"
            }
          ],
          "total": 1,
          "page": 1,
          "size": 20
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "description": "Spring ProblemDetail；字段可能随框架渲染上下文增减。",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Forbidden"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599,
            "example": 403
          },
          "detail": {
            "type": "string",
            "description": "可能缺省。"
          },
          "instance": {
            "type": "string",
            "description": "可能缺省；通常为请求路径。",
            "format": "uri-reference"
          }
        },
        "additionalProperties": true
      },
      "BootError": {
        "type": "object",
        "description": "未被 ApiErrors 捕获的 ResponseStatusException 可能走 Boot /error。当前没有启用 spring.mvc.problemdetails.enabled，不能保证这些错误全部为 ProblemDetail。",
        "required": [
          "status",
          "error"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "path": {
            "type": "string",
            "example": "/auth/api/admin/clients"
          },
          "message": {
            "type": "string",
            "description": "默认配置通常不暴露 reason/message。"
          }
        },
        "additionalProperties": true
      },
...
    },
    "examples": {
      "CsrfToken": {
        "summary": "会话绑定的 CSRF 令牌",
        "value": {
          "headerName": "X-CSRF-TOKEN",
          "parameterName": "_csrf",
          "token": "csrf-token-placeholder"
        },
        "description": "保存同一响应会话 Cookie；此 token 仅为占位值。"
      },
      "RegisterAccount": {
        "summary": "注册普通账号",
        "value": {
          "username": "Alice",
          "password": "example-passphrase-2026"
        }
      },
      "ChangePassword": {
        "summary": "修改当前密码",
        "value": {
          "currentPassword": "example-passphrase-2026",
          "newPassword": "replacement-passphrase-2026"
        }
      },
      "AccountEnabled": {
        "summary": "已启用账号",
        "value": {
          "id": 42,
          "subject": "7d8a8d55-e9cb-4db4-a945-6f4c21cbab31",
          "username": "alice",
          "enabled": true,
          "createTime": "2026-10-03T02:00:00",
          "updateTime": "2026-10-03T02:00:00"
        }
      },
...
      "Enable": {
        "summary": "启用",
        "value": {
          "enabled": true
        }
      },
...
      "AccountsPage": {
        "summary": "账号分页",
        "value": {
          "items": [
            {
              "id": 42,
              "subject": "7d8a8d55-e9cb-4db4-a945-6f4c21cbab31",
              "username": "alice",
              "enabled": true,
              "createTime": "2026-10-03T02:00:00",
              "updateTime": "2026-10-03T02:00:00"
            }
          ],
          "total": 1,
          "page": 1,
          "size": 20
        }
      },
...
    }
  },
...
}